Added functionality to post comments on articles, all tests passing

+++ models/articles.models.js
@@ -2,7 +2,7 @@ const db = require("../db/connection");
 
 function fetchArticles() {
     return db.query(`
-        SELECT title, articles.article_id, topic, articles.author, articles.created_at, articles.votes, article_img_url, COUNT(comments.article_id) AS "comment_count" 
+        SELECT title, articles.article_id, topic, articles.author, articles.created_at, articles.votes, article_img_url, COUNT(comments.article_id)::INT AS "comment_count" 
         FROM articles LEFT JOIN comments ON articles.article_id = comments.article_id 
         GROUP BY articles.article_id 
         ORDER BY created_at DESC;
